Fixed enemies bugs

+++ NONDA/Assets/_Scripts/Model/Quiz.cs
@@ -2,5 +2,5 @@
 public class Quiz {
 	public string question;
 	public bool isTrue;
-	public float questionValue;
+	public int questionValue;
 }
